pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "semantic-pdf-rag"
version = "0.1.0"
description = "CLI-first semantic PDF RAG system with Clean Architecture"
readme = "readme.md"
requires-python = ">=3.9"
dependencies = [
  "langchain",
  "langchain-postgres",
  "langchain-openai",
  "langchain-google-genai",
  "openai",
  "google-generativeai",
  "psycopg2-binary",
  "pgvector",
  "sqlalchemy",
  "pydantic",
  "pydantic-settings",
  "python-dotenv",
  "typer",
  "pypdf",
]

[project.optional-dependencies]
dev = [
  "pytest",
]

[tool.pytest.ini_options]
testpaths = ["tests"]
pythonpath = ["."]
